Fix EarlyStoppingCallback error in train-codebert notebook
This PR addresses the error that occurs during training in the train-codebert.ipynb notebook:
```
Error during training: EarlyStoppingCallback requires metric_for_best_model to be defined
```

The issue was that even though we removed the `metric_for_best_model` parameter from `TrainingArguments` in our previous fix, we were still using the `EarlyStoppingCallback` which requires this parameter to function.

Changes made:
1. Completely removed the `EarlyStoppingCallback` from the Trainer initialization
2. Simplified the code by removing the nested try-except block that was used to handle EarlyStoppingCallback errors
3. Added clear comments explaining why the callback was removed

This fix ensures that the notebook will run without errors when using an older version of the transformers library that doesn't support all the newer parameters.